test_worker_python_requirements.txt
python-dotenv 
jupyterlab 
ipywidgets
pandas 
numpy 
matplotlib